Feat(WPM-20): Update audio backend (#50)
Accept test script
Accept moving unity to global vendors
Accept non main files
Accept test's CMakeLists
Accept main library audio backend CMake
Accept main library file of audio backend
Accept main library header of audio backend
---
Downgrade CMake minimum required version to 3.21
Enable AAUDIO for android
Windows related update
---
Update pubspec version
Update pubspec.lock

+++ native_library/wp_player_lib/wp_playerlib_id.c
@@ -1,12 +1,11 @@
 #include <stdlib.h>
 #include <time.h>
 
-
 uint32_t wp_playerlib_id_next(void) {
-    static int seeded = 0;
-    if (!seeded) {
-        srand(time(NULL));
-        seeded = 1;
-    }
-    return (uint32_t)rand();
+  static int seeded = 0;
+  if (!seeded) {
+    srand((unsigned int)time(NULL));
+    seeded = 1;
+  }
+  return (uint32_t)rand();
 }
